YUNKE-3283 【Server】作为云视课堂运营者，我希望所有学校能在一套系统中运行，而不是每个学校一套部署

+++ field_int64.go
@@ -9,7 +9,7 @@ type Int64Field struct {
 }
 
 // NewInt64Field 创建一个整形字段
-func NewInt64Field(key string, value int) *Int64Field {
+func NewInt64Field(key string, value int64) *Int64Field {
 	return &Int64Field{
 		StringKey: StringKey{
 			key: key,
